COMMONLUA: fixed float vec3 issue

+++ src/modules/commonlua/LUAFunctions.h
@@ -175,7 +175,7 @@ struct clua_vecnew {};
 template<int N>
 struct clua_vecnew<glm::vec<N, float> > {
 static int vecnew(lua_State* s) {
-	glm::vec<N, int> array;
+	glm::vec<N, float> array;
 	float value = 0.0f;
 	for (size_t i = 0; i < sizeof(array) / sizeof(array[0]); ++i) {
 		array[i] = luaL_optnumber(s, i + 1, value);
